fix: Shape.color setter left an all-integer input as int64
color=[1, 0, 0, 1] (a natural way to write opaque red) stayed int64 --
nothing needs 0-255 normalisation when every value is <= 1, so the
only thing that made a 3-length int color safe was concatenate()'s
dtype promotion when appending the alpha default; a fully-specified
4-length int color skipped that entirely. self.color[3] (opacity)
being numpy.int64 rather than float breaks real (non-mocked)
json.dumps() sends in swift's SwiftRoute.py -- TypeError: Object of
type int64 is not JSON serializable.

Found live-testing spatialgeometry.Path in swift, unrelated to Path
itself. Never caught by the existing test suite because swift's own
protocol tests are entirely FakeBrowser-mocked -- no real json.dumps()
ever runs in that path. Verified fixed end-to-end with a real swift
browser round-trip, not just the new unit test.

+++ src/spatialgeometry/geom/Shape.py
@@ -252,7 +252,16 @@ class Shape(SceneNode):
         elif value is None:
             value = default_color
         else:
-            value = array(value)
+            # dtype=float forced explicitly -- an all-integer input (e.g.
+            # the very natural color=[1, 0, 0, 1] for opaque red) would
+            # otherwise stay int64 whenever nothing needs 0-255
+            # normalisation below, and self.color[3] (opacity) being a
+            # numpy.int64 rather than a float breaks real (non-mocked)
+            # json.dumps() sends in SwiftRoute.py -- TypeError: Object of
+            # type int64 is not JSON serializable. Never caught by the
+            # existing protocol tests since they're FakeBrowser-mocked and
+            # never actually round-trip through json.dumps().
+            value = array(value, dtype=float)
 
             if any(value > 1.0):
                 value = value / 255.0
